add magit-find-file-ignore-extensions
replacing `magit-find-file-skip-images' which used a constant list
of a small subset of image file extensions.  Even if that list were
"complete" this option would still be problematic because it is it
provides very arbitrary filtering; only a subset of potential users
of this package would want to filter images, while others might want
to ignore other kinds files.  For that reason use the empty list as
default value of the new option.